TST fixed tests
